backend/requirements.txt
# Django
Django==1.11.5
django-extensions==1.9.1

# Rest framwork
djangorestframework==3.6.4
djangorestframework_jwt==1.11.0
django-cors-headers==2.1.0